Refactor: move enum class "CommandType" to "common" package
