+++ src/main/java/data/common/CommandType.java
@@ -1,4 +1,4 @@
-package commands;
+package data.common;
 
 import data.exceptions.BarryException;
